requirements.txt
Django==2.1.5
martor==1.3.5
sentry-sdk==0.6.9
gunicorn==19.9.0
psycopg2-binary==2.7.7
django-userena-ce==4.1.0
